chore: moved cleanup to separate monobehavior
