+++ Runtime/Core/Scripts/Utils/DestroyMesh.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9d21b91e22ee6f548b51e598aacec332
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
